+++ init/yyt/test/hello-world.js
@@ -1,0 +1,8 @@
+module.exports = {
+  // '@disable': true,
+  'hello-world-test': function (client) {
+    return client
+      .checkPageError('http://www.yy.com')
+      .end();
+  }
+};