+++ test/config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  nightwatch: {
-    src_folders: ['uitest'],
-    output_folder: './reports',
-    html_report_folder: './reports'
-  }
-};